# fstickbackend/registry-service/src/main/java/ru/fstick/registry_service/dto/api/view/PluginViewExtend.java
package ru.fstick.registry_service.dto.api.view;

import lombok.Builder;
import lombok.Data;

import java.util.List;
import java.util.UUID;
@Data
@Builder
public class PluginViewExtend {
    private UUID id;
    private UUID authorId;
    private String name;
    private String description;
    private String category;
    private List<String> tags;
    private String status;
    private String iconUrl;
    private String createdAt;
    private String updatedAt;
    private List<BranchView> branches;
    private LastRejectionView lastRejection;
    private List<ScreenshotView> screenshots;
}